1.9
modificaciones varias

+++ Veterinaria/src/veterinaria/AccesoADatos/ClienteData.java
@@ -95,7 +95,7 @@ public class ClienteData {
                 cliente.setApellido(rs.getString("apellido"));
                 cliente.setNombre(rs.getString("nombre"));
                 cliente.setDireccion(rs.getString("direccion"));
-                cliente.setTelefono(rs.getInt("telefono"));
+                cliente.setTelefono(rs.getLong("telefono"));
                 cliente.setPersonaAlternativa(rs.getString("personaAlternativa"));
                 cliente.setEstado(rs.getBoolean("estado"));
                 lista.add(cliente);
